# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Content Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}/scenes:
    get:
      operationId: getProjectScenes
      tags: [scenes]
      summary: Get scenes for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scenes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scene"
    post:
      operationId: generateScenes
      tags: [scenes]
      summary: Generate scenes for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SceneGenerateInput"
      responses:
        "200":
          description: Generated scenes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scene"
...
components:
  schemas:
...
    Scene:
      type: object
      required: [id, projectId, order, text, voiceScript, duration]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        order:
          type: integer
        text:
          type: string
        voiceScript:
          type: string
        cta:
          type: ["string", "null"]
        duration:
          type: number
        animationPreset:
          type: ["string", "null"]
        backgroundType:
          type: ["string", "null"]
        subtitleMode:
          type: ["string", "null"]
    SceneGenerateInput:
      type: object
      required: [count]
      properties:
        count:
          type: integer
          minimum: 1
          maximum: 20
        topic:
          type: string
        style:
          type: string
...